src/requirements.txt
llama-cpp-python
langchain
psutil